update field type

+++ database/migrations/2023_05_20_183957_create_device_data_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('device_data', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->foreignId('device_id')->constrained('devices')->cascadeOnDelete()->cascadeOnUpdate();
-            $table->unsignedTinyInteger('filled');
-            $table->unsignedTinyInteger('unfilled');
+            $table->double('filled');
+            $table->double('unfilled');
             $table->timestamps();
         });
     }
